+++ examples/dynamicExample.js
@@ -14,8 +14,10 @@ const pool = new DynamicThreadPool(
   }
 )
 let poolFull = 0
+let poolReady = 0
 let poolBusy = 0
 pool.emitter.on(PoolEvents.full, () => poolFull++)
+pool.emitter.on(PoolEvents.ready, () => poolReady++)
 pool.emitter.on(PoolEvents.busy, () => poolBusy++)
 
 let resolved = 0
@@ -29,6 +31,7 @@ for (let i = 1; i <= iterations; i++) {
       if (resolved === iterations) {
         console.info('Time taken is ' + (performance.now() - start))
         console.info('The pool was full for ' + poolFull + ' times')
+        console.info('The pool was ready for ' + poolReady + ' times')
         return console.info('The pool was busy for ' + poolBusy + ' times')
       }
       return null