+++ src/pools/worker-node.ts
@@ -75,7 +75,7 @@ implements IWorkerNode<Worker, Data> {
       id: this.getWorkerId(worker, workerType),
       type: workerType,
       dynamic: false,
-      ready: true
+      ready: false
     }
   }